api/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
chevron==0.14.0
click==8.1.3
comm==0.1.3
crcmod==1.7
cryptography==41.0.1
dataclasses-json==0.6.1
debugpy==1.6.7
decorator==5.1.1
distro==1.8.0
dnspython==2.4.2
environs==9.5.0
executing==1.2.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Markdown==0.2
Flask-SQLAlchemy==3.0.5
frozenlist==1.3.3
grpcio==1.56.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.24.0
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.327
langfuse==2.4.2
langsmith==0.0.56
lazy-object-proxy==1.9.0
litereval==0.0.11
markdown2==2.4.10
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==10.1.0
motor==3.3.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.25.2
openai==1.6.1
openapi-core==0.18.1
openapi-schema-validator==0.6.2
openapi-spec-validator==0.6.0
oss2==2.18.1
packaging==23.1
pandas==2.0.3
parse==1.19.1
parso==0.8.3
pathable==0.4.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.1
prance==23.6.21.0
prompt-toolkit==3.0.39
protobuf==4.24.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.13
pyee==8.2.2
Pygments==2.15.1
PyJWT==2.7.0
pymilvus==2.2.15
pymongo==4.5.0
PyMySQL==1.1.0
pyppdf==0.1.2
pyppeteer==1.0.2
pysandbox==66.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
redis==4.6.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rpds-py==0.10.6
ruamel.yaml==0.18.3
ruamel.yaml.clib==0.2.8
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
stack-data==0.6.2
tenacity==8.2.3
tiktoken==0.4.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.15
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.3.3
wrapt==1.14.0
yarl==1.9.2
zipp==3.16.2
